+++ lib/standard_playing.rb
@@ -3,6 +3,9 @@ require_relative 'deck'
 
 class StandardPlaying < Deck
 
+  VALUES = Card::FACECARDS.keys + (2..10).map(&:to_s).to_a
+  SUITS = Card::SUITDEC.keys
+  
   def initialize()
     @cards = []
     for v in 2..14
